+++ src/index.ts
@@ -3,6 +3,7 @@ import {defineIntlDisplayNamesElements} from './elements/displaynames';
 import {defineIntlListFormatElements} from './elements/listformat';
 import {defineIntlLocaleElements} from './elements/locale';
 import {defineIntlNumberFormatElements} from './elements/numberformat';
+import {defineIntlPluralRulesElements} from './elements/pluralrules';
 import {defineIntlRelativeTimeFormatElements} from './elements/relativetimeformat';
 import {defineIntlSegmenterElements} from './elements/segmenter';
 
@@ -12,6 +13,7 @@ export function defineIntlElements() {
   defineIntlListFormatElements();
   defineIntlLocaleElements();
   defineIntlNumberFormatElements();
+  defineIntlPluralRulesElements();
   defineIntlRelativeTimeFormatElements();
   defineIntlSegmenterElements();
 }
@@ -21,5 +23,6 @@ export * from './elements/displaynames';
 export * from './elements/listformat';
 export * from './elements/locale';
 export * from './elements/numberformat';
+export * from './elements/pluralrules';
 export * from './elements/relativetimeformat';
 export * from './elements/segmenter';